fix: import button now confirms whether you want to overwrite pins style: updating button styling

+++ src/components/UserPins.tsx
@@ -133,14 +133,14 @@ export function ExportUserPinsButton() {
     const pins_json_file = new Blob([pins_json], { type: "application/json" });
 
     return (
-        <button>
+        <button className="bg-blue-900 w-full outline outline-1 p-1">
             <a
                 download="user_pins.json"
                 target="_blank"
                 rel="noreferrer"
                 href={URL.createObjectURL(pins_json_file)}
             >
-                Export pins
+                Export Pins
             </a>
         </button>
     );
@@ -152,22 +152,27 @@ export function ImportUserPinsButton({
     setUserPins: React.Dispatch<React.SetStateAction<LocalStoragePin[]>>
 }) {
     return (
-        <label htmlFor="upload" className="cursor-pointer">
-            <span>Import pins</span>
+        <label
+            htmlFor="upload"
+            className="flex justify-center items-center w-full cursor-pointer bg-blue-900 outline outline-1 p-1 text-center"
+        >
+            <span>Import Pins</span>
             <input
                 type="file"
                 accept=".json"
                 id="upload"
                 className="hidden"
                 onChange={(event) => {
-                    const file = event.target.files[0];
+                    const file = (event.target.files ?? [])[0];
                     if (!file) return;
 
                     const reader = new FileReader();
                     reader.onload = (e) => {
-                        const pins_json = e.target?.result;
-                        localStorage.setItem("user_pins", pins_json);
-                        setUserPins(JSON.parse(pins_json));
+                        const pins_json = e.target?.result as string;
+                        if (window.confirm("This will overwrite your current pins. Are you sure you want to continue?")) {
+                            localStorage.setItem("user_pins", pins_json);
+                            setUserPins(JSON.parse(pins_json));
+                        }
                     };
                     reader.onerror = () => {
                         console.error("error: failed to read user pins");
@@ -186,11 +191,13 @@ export function ClearUserPinsButton({
     setUserPins: React.Dispatch<React.SetStateAction<LocalStoragePin[]>>
 }) {
     return (
-        <button onClick={() => {
-            setUserPins([]);
-            localStorage.setItem("user_pins", JSON.stringify([]));
-        }}>
-            Clear pins
+        <button
+            className="bg-red-600 p-1 w-full outline outline-1"
+            onClick={() => {
+                setUserPins([]);
+                localStorage.setItem("user_pins", JSON.stringify([]));
+            }}>
+            Clear Pins
         </button>
     );
 }
